Fix adjustment window to cover every past year, not just one
getAdjustmentWindowData only ever fetched the single "live" target year,
so every other year's tab had nothing to render regardless of what was
actually logged for it. Now fetches one panel's worth of data per year
that has actually reached its own Dec25-Jan31 window (excludes the
current, still-being-read year, which hasn't reached its window at all
yet and shouldn't be mislabeled "closed"). Also fixes a latent crash in
both this panel and the book dossier's adjustment-history section: their
queries read numeric rank/score columns without a float8 cast, so pg
returned them as strings and .toFixed() blew up the moment a score-kind
adjustment actually had a reason attached.

+++ app/books/[id]/page.tsx
@@ -121,12 +121,12 @@ type AdjustmentHistoryEvent = {
 // doesn't show up here at all. See app/rankings/adjustmentWindow.ts.
 async function getAdjustmentHistory(bookId: number): Promise<AdjustmentHistoryEvent[]> {
   const { rows } = await pool.query<AdjustmentHistoryEvent>(
-    `select 'rank' as kind, rc.old_rank::numeric as old_val, rc.new_rank::numeric as new_val,
+    `select 'rank' as kind, rc.old_rank::float8 as old_val, rc.new_rank::float8 as new_val,
             rc.reason, to_char(rc.changed_at, 'YYYY-MM-DD"T"HH24:MI:SS') as changed_at
      from rank_changes rc
      where rc.book_id = $1 and rc.reason is not null
      union all
-     select 'score' as kind, sc.old_score as old_val, sc.new_score as new_val,
+     select 'score' as kind, sc.old_score::float8 as old_val, sc.new_score::float8 as new_val,
             sc.reason, to_char(sc.changed_at, 'YYYY-MM-DD"T"HH24:MI:SS') as changed_at
      from score_changes sc
      where sc.book_id = $1 and sc.reason is not null
